Tester updates- created base class for modularity between j750 and uflex. Added uflex tester files
