+++ rust/origen/src/testers/igxl/mod.rs
@@ -1,4 +1,7 @@
+mod base;
 pub mod j750;
 pub mod ultraflex;
 
+pub use j750::J750;
 pub use ultraflex::UltraFlex;
+pub use base::IGXLBase;
