+++ rust/origen/src/testers/igxl/ultraflex/mod.rs
@@ -1,20 +1,33 @@
-use crate::core::tester::{Interceptor, TesterAPI, TesterID};
+use crate::core::tester::{Interceptor, TesterID};
 use crate::testers::SupportedTester;
 use origen_metal::prog_gen::SupportedTester as ProgGenSupportedTester;
-//use crate::generator::ast::Node;
-//use crate::Result;
-//use std::path::PathBuf;
+use super::base::IGXLBase;
 
+
+/// Teradyne UltraFlex tester implementation.
+/// 
+/// Inherits IGXL pattern generation from IGXLBase trait.
+/// No tester-specific customization currently required.
 #[derive(Debug, Clone)]
 pub struct UltraFlex {}
 
-impl Default for UltraFlex {
-    fn default() -> Self {
+impl UltraFlex {
+    pub fn default() -> Self {
         Self {}
     }
 }
 
+impl std::default::Default for UltraFlex {
+    fn default() -> Self {
+        Self::default()
+    }
+}
+
+/// Uses default interceptor behavior (no command interception).
 impl Interceptor for UltraFlex {}
+
+/// Provides IGXL pattern generation via blanket impl in base.rs.
+impl IGXLBase for UltraFlex {}
 
 impl TesterID for UltraFlex {
     fn id(&self) -> SupportedTester {
@@ -25,5 +38,3 @@ impl TesterID for UltraFlex {
         ProgGenSupportedTester::ULTRAFLEX
     }
 }
-
-impl TesterAPI for UltraFlex {}
